+++ src/common/private.h
@@ -185,7 +185,7 @@ extern const struct kqueue_vtable kqops;
 /*
  * knote internal API
  */
-struct knote * knote_lookup(struct filter *, short);
+struct knote * knote_lookup(struct filter *, uintptr_t);
 //DEADWOOD: struct knote * knote_get_by_data(struct filter *filt, intptr_t);
 struct knote * knote_new(void);
 #define knote_retain(kn) atomic_inc(&kn->kn_ref)
